safari-600.7-branch/LayoutTests/platform/ios-sim/accessibility/identifier-expected.txt

button 1
This tests that id maps to AXIdentifier correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button1.iphoneIdentifier is 'AXIdentifier: button1'
PASS successfullyParsed is true

TEST COMPLETE
